Split off DateInfoPanel and ParkInfoPanel from GameBottomToolbar
