+++ src/openrct2-ui/interface/Theme.cpp
@@ -159,6 +159,8 @@ namespace OpenRCT2::Ui
         // WindowClass                              WindowClassSZ                        WindowName                                        windowColours
         { WindowClass::topToolbar,                  "WC_TOP_TOOLBAR",                    STR_THEMES_WINDOW_TOP_TOOLBAR,                    { opaque(Drawing::Colour::lightBlue),               opaque(Drawing::Colour::darkGreen),               opaque(Drawing::Colour::darkBrown),             opaque(Drawing::Colour::grey)         } },
         { WindowClass::bottomToolbar,               "WC_BOTTOM_TOOLBAR",                 STR_THEMES_WINDOW_BOTTOM_TOOLBAR,                 { translucent(Drawing::Colour::darkGreen),          opaque(Drawing::Colour::black),                    opaque(Drawing::Colour::black),                  opaque(Drawing::Colour::brightGreen) } },
+        { WindowClass::parkInfoPanel,               "WC_PARK_INFO_PANEL",                STR_THEMES_WINDOW_BOTTOM_TOOLBAR,                 { translucent(Drawing::Colour::darkGreen),          opaque(Drawing::Colour::brightGreen)                                                                                                     } },
+        { WindowClass::dateInfoPanel,               "WC_DATE_INFO_PANEL",                STR_THEMES_WINDOW_BOTTOM_TOOLBAR,                 { translucent(Drawing::Colour::darkGreen)                                                                                                                                                    } },
         { WindowClass::ride,                        "WC_RIDE",                           STR_THEMES_WINDOW_RIDE,                           { opaque(Drawing::Colour::grey),                     opaque(Drawing::Colour::bordeauxRed),             opaque(Drawing::Colour::saturatedGreen)                                     } },
         { WindowClass::rideConstruction,            "WC_RIDE_CONSTRUCTION",              STR_THEMES_WINDOW_RIDE_CONSTRUCTION,              { opaque(Drawing::Colour::darkBrown),               opaque(Drawing::Colour::darkBrown),               opaque(Drawing::Colour::darkBrown)                                          } },
         { WindowClass::rideList,                    "WC_RIDE_LIST",                      STR_THEMES_WINDOW_RIDE_LIST,                      { opaque(Drawing::Colour::grey),                     opaque(Drawing::Colour::bordeauxRed),             opaque(Drawing::Colour::bordeauxRed)                                        } },
@@ -232,7 +234,9 @@ namespace OpenRCT2::Ui
 
     static constexpr std::array kPredefinedThemeRCT1Entries = std::to_array<UIThemeWindowEntry>({
         { WindowClass::topToolbar,             { opaque(Drawing::Colour::grey),             opaque(Drawing::Colour::grey),             opaque(Drawing::Colour::grey),                opaque(Drawing::Colour::grey),     opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black) } },
-        { WindowClass::bottomToolbar,          { translucent(Drawing::Colour::grey),        opaque(Drawing::Colour::grey),             opaque(Drawing::Colour::voidBackground),                opaque(Drawing::Colour::yellow),   opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black) } },
+        { WindowClass::bottomToolbar,          { translucent(Drawing::Colour::grey),        opaque(Drawing::Colour::grey),             opaque(Drawing::Colour::voidBackground),      opaque(Drawing::Colour::yellow),   opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black) } },
+        { WindowClass::parkInfoPanel,          { translucent(Drawing::Colour::grey),        opaque(Drawing::Colour::yellow) } },
+        { WindowClass::dateInfoPanel,          { translucent(Drawing::Colour::grey) } },
         { WindowClass::ride,                   { opaque(Drawing::Colour::bordeauxRed),     opaque(Drawing::Colour::grey),             opaque(Drawing::Colour::saturatedGreen),     opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black) } },
         { WindowClass::rideList,               { opaque(Drawing::Colour::bordeauxRed),     opaque(Drawing::Colour::grey),             opaque(Drawing::Colour::grey),                opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black) } },
         { WindowClass::constructRide,          { opaque(Drawing::Colour::bordeauxRed),     opaque(Drawing::Colour::grey),             opaque(Drawing::Colour::grey),                opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black),    opaque(Drawing::Colour::black) } },
