+++ src/openrct2-ui/windows/GameBottomToolbar.cpp
@@ -43,39 +43,19 @@ namespace OpenRCT2::Ui::Windows
 {
     enum WindowGameBottomToolbarWidgetIdx : WidgetIndex
     {
-        WIDX_LEFT_OUTSET,
-        WIDX_LEFT_INSET,
-        WIDX_MONEY,
-        WIDX_GUESTS,
-        WIDX_PARK_RATING,
-
         WIDX_MIDDLE_OUTSET,
         WIDX_MIDDLE_INSET,
         WIDX_NEWS_SUBJECT,
         WIDX_NEWS_LOCATE,
-
-        WIDX_RIGHT_OUTSET,
-        WIDX_RIGHT_INSET,
-        WIDX_DATE
     };
 
     // clang-format off
-    static constexpr Widget window_game_bottom_toolbar_widgets[] =
+    static constexpr Widget kGameBottomToolbarWidgets[] =
     {
-        makeWidget({  0,  0}, {142, 34}, WidgetType::imgBtn,      WindowColour::primary                                                     ), // Left outset panel
-        makeWidget({  2,  2}, {138, 30}, WidgetType::empty,       WindowColour::primary                                                     ), // Left inset panel
-        makeWidget({  2,  1}, {138, 12}, WidgetType::hiddenButton,WindowColour::primary , 0xFFFFFFFF, STR_PROFIT_PER_WEEK_AND_PARK_VALUE_TIP), // Money window
-        makeWidget({  2, 11}, {138, 12}, WidgetType::hiddenButton,WindowColour::primary                                                     ), // Guests window
-        makeWidget({  2, 21}, {138, 11}, WidgetType::hiddenButton,WindowColour::primary , 0xFFFFFFFF, STR_PARK_RATING_TIP                   ), // Park rating window
-
         makeWidget({142,  0}, {356, 34}, WidgetType::imgBtn,      WindowColour::tertiary                                                    ), // Middle outset panel
         makeWidget({144,  2}, {352, 30}, WidgetType::hiddenButton,WindowColour::tertiary                                                    ), // Middle inset panel
         makeWidget({147,  5}, { 24, 24}, WidgetType::flatBtn,     WindowColour::secondary, 0xFFFFFFFF, STR_SHOW_SUBJECT_TIP                 ), // Associated news item window
-        makeWidget({469,  5}, { 24, 24}, WidgetType::flatBtn,     WindowColour::secondary, ImageId(SPR_LOCATE), STR_LOCATE_SUBJECT_TIP      ), // Scroll to news item target
-
-        makeWidget({498,  0}, {142, 34}, WidgetType::imgBtn,      WindowColour::primary                                                     ), // Right outset panel
-        makeWidget({500,  2}, {138, 30}, WidgetType::empty,       WindowColour::primary                                                     ), // Right inset panel
-        makeWidget({500,  2}, {138, 12}, WidgetType::hiddenButton,WindowColour::primary                                                     ), // Date
+        makeWidget({469,  5}, { 24, 24}, WidgetType::flatBtn,     WindowColour::secondary, ImageId(SPR_LOCATE), STR_LOCATE_SUBJECT_TIP      )  // Scroll to news item target
     };
     // clang-format on
 
@@ -84,163 +64,6 @@ namespace OpenRCT2::Ui::Windows
     class GameBottomToolbar final : public Window
     {
     private:
-        Drawing::Colour GetHoverWidgetColour(WidgetIndex index)
-        {
-            return (
-                gHoverWidget.windowClassification == WindowClass::bottomToolbar && gHoverWidget.widgetIndex == index
-                    ? static_cast<Drawing::Colour>(Drawing::Colour::white)
-                    : colours[0].colour);
-        }
-
-        void DrawLeftPanel(RenderTarget& rt)
-        {
-            const auto& leftPanelWidget = widgets[WIDX_LEFT_OUTSET];
-
-            const auto topLeft = windowPos + ScreenCoordsXY{ leftPanelWidget.left + 1, leftPanelWidget.top + 1 };
-            const auto bottomRight = windowPos + ScreenCoordsXY{ leftPanelWidget.right - 1, leftPanelWidget.bottom - 1 };
-            // Draw green inset rectangle on panel
-            Rectangle::fillInset(
-                rt, { topLeft, bottomRight }, colours[0], Rectangle::BorderStyle::inset, Rectangle::FillBrightness::light,
-                Rectangle::FillMode::none);
-
-            // Figure out how much line height we have to work with.
-            uint32_t line_height = FontGetLineHeight(FontStyle::medium);
-
-            auto& gameState = getGameState();
-
-            // Draw money
-            if (!gameState.park.flags.has(ParkFlag::noMoney))
-            {
-                const auto& widget = widgets[WIDX_MONEY];
-                auto screenCoords = ScreenCoordsXY{ windowPos.x + widget.midX(),
-                                                    windowPos.y + widget.midY() - (line_height == 10 ? 5 : 6) };
-
-                auto colour = GetHoverWidgetColour(WIDX_MONEY);
-                StringId stringId = gameState.park.cash < 0 ? STR_BOTTOM_TOOLBAR_CASH_NEGATIVE : STR_BOTTOM_TOOLBAR_CASH;
-                auto ft = Formatter();
-                ft.Add<money64>(gameState.park.cash);
-                drawText(rt, screenCoords, stringId, ft, { colour, TextAlignment::centre });
-            }
-
-            static constexpr StringId _guestCountFormats[] = {
-                STR_BOTTOM_TOOLBAR_NUM_GUESTS_STABLE,
-                STR_BOTTOM_TOOLBAR_NUM_GUESTS_DECREASE,
-                STR_BOTTOM_TOOLBAR_NUM_GUESTS_INCREASE,
-            };
-
-            static constexpr StringId _guestCountFormatsSingular[] = {
-                STR_BOTTOM_TOOLBAR_NUM_GUESTS_STABLE_SINGULAR,
-                STR_BOTTOM_TOOLBAR_NUM_GUESTS_DECREASE_SINGULAR,
-                STR_BOTTOM_TOOLBAR_NUM_GUESTS_INCREASE_SINGULAR,
-            };
-
-            // Draw guests
-            {
-                const auto& widget = widgets[WIDX_GUESTS];
-                auto screenCoords = ScreenCoordsXY{ windowPos.x + widget.midX(), windowPos.y + widget.midY() - 6 };
-
-                StringId stringId = gameState.park.numGuestsInPark == 1
-                    ? _guestCountFormatsSingular[gameState.park.guestChangeModifier]
-                    : _guestCountFormats[gameState.park.guestChangeModifier];
-                auto colour = GetHoverWidgetColour(WIDX_GUESTS);
-                auto ft = Formatter();
-                ft.Add<uint32_t>(gameState.park.numGuestsInPark);
-                drawText(rt, screenCoords, stringId, ft, { colour, TextAlignment::centre });
-            }
-
-            // Draw park rating
-            {
-                const auto& widget = widgets[WIDX_PARK_RATING];
-                auto screenCoords = windowPos + ScreenCoordsXY{ widget.left + 11, widget.midY() - 5 };
-
-                DrawParkRating(
-                    rt, colours[3].colour, false, screenCoords, std::max(10, ((gameState.park.rating / 4) * 263) / 256));
-            }
-        }
-
-        void DrawParkRating(RenderTarget& rt, Colour colour, bool blink, const ScreenCoordsXY& coords, uint8_t factor)
-        {
-            int16_t bar_width = (factor * 114) / 255;
-            Rectangle::fillInset(
-                rt, { coords + ScreenCoordsXY{ 1, 1 }, coords + ScreenCoordsXY{ 114, 9 } }, colours[0],
-                Rectangle::BorderStyle::inset, Rectangle::FillBrightness::light, Rectangle::FillMode::none);
-            if (!blink || GameIsPaused() || (gCurrentRealTimeTicks & 8))
-            {
-                if (bar_width > 2)
-                {
-                    Rectangle::fillInset(
-                        rt, { coords + ScreenCoordsXY{ 2, 2 }, coords + ScreenCoordsXY{ bar_width - 1, 8 } },
-                        ColourWithFlags{ colour });
-                }
-            }
-
-            // Draw thumbs on the sides
-            GfxDrawSprite(rt, ImageId(SPR_RATING_LOW), coords - ScreenCoordsXY{ 14, 0 });
-            GfxDrawSprite(rt, ImageId(SPR_RATING_HIGH), coords + ScreenCoordsXY{ 114, 0 });
-        }
-
-        void DrawRightPanel(RenderTarget& rt)
-        {
-            const auto& rightPanelWidget = widgets[WIDX_RIGHT_OUTSET];
-
-            const auto topLeft = windowPos + ScreenCoordsXY{ rightPanelWidget.left + 1, rightPanelWidget.top + 1 };
-            const auto bottomRight = windowPos + ScreenCoordsXY{ rightPanelWidget.right - 1, rightPanelWidget.bottom - 1 };
-            // Draw green inset rectangle on panel
-            Rectangle::fillInset(
-                rt, { topLeft, bottomRight }, colours[0], Rectangle::BorderStyle::inset, Rectangle::FillBrightness::light,
-                Rectangle::FillMode::none);
-
-            auto screenCoords = ScreenCoordsXY{ (rightPanelWidget.left + rightPanelWidget.right) / 2 + windowPos.x,
-                                                rightPanelWidget.top + windowPos.y + 2 };
-
-            // Date
-            auto& date = GetDate();
-            int32_t year = date.GetYear() + 1;
-            int32_t month = date.GetMonth();
-            int32_t day = date.GetDay();
-
-            auto colour = GetHoverWidgetColour(WIDX_DATE);
-            StringId stringId = DateFormatStringFormatIds[Config::Get().general.dateFormat];
-            auto ft = Formatter();
-            ft.Add<StringId>(DateDayNames[day]);
-            ft.Add<int16_t>(month);
-            ft.Add<int16_t>(year);
-            drawText(rt, screenCoords, stringId, ft, { colour, TextAlignment::centre });
-
-            // Figure out how much line height we have to work with.
-            uint32_t line_height = FontGetLineHeight(FontStyle::medium);
-
-            // Temperature
-            screenCoords = { windowPos.x + rightPanelWidget.left + 15, static_cast<int32_t>(screenCoords.y + line_height + 1) };
-
-            int32_t temperature = getGameState().weatherCurrent.temperature;
-            StringId format = STR_CELSIUS_VALUE;
-            if (Config::Get().general.temperatureFormat == TemperatureUnit::fahrenheit)
-            {
-                temperature = Weather::celsiusToFahrenheit(temperature);
-                format = STR_FAHRENHEIT_VALUE;
-            }
-            ft = Formatter();
-            ft.Add<int16_t>(temperature);
-            drawText(rt, screenCoords + ScreenCoordsXY{ 0, 6 }, format, ft);
-            screenCoords.x += 30;
-
-            // Current weather
-            auto currentWeatherSpriteId = Weather::getWeatherSpriteId(getGameState().weatherCurrent.weatherType);
-            GfxDrawSprite(rt, ImageId(currentWeatherSpriteId), screenCoords);
-
-            // Next weather
-            auto nextWeatherSpriteId = Weather::getWeatherSpriteId(getGameState().weatherNext.weatherType);
-            if (currentWeatherSpriteId != nextWeatherSpriteId)
-            {
-                if (getGameState().weatherUpdateTimer < 960)
-                {
-                    GfxDrawSprite(rt, ImageId(SPR_NEXT_WEATHER), screenCoords + ScreenCoordsXY{ 27, 5 });
-                    GfxDrawSprite(rt, ImageId(nextWeatherSpriteId), screenCoords + ScreenCoordsXY{ 40, 0 });
-                }
-            }
-        }
-
         void DrawNewsItem(RenderTarget& rt)
         {
             const auto& middleOutsetWidget = widgets[WIDX_MIDDLE_OUTSET];
@@ -391,43 +214,10 @@ namespace OpenRCT2::Ui::Windows
             }
         }
 
-        void InvalidateDirtyWidgets()
-        {
-            if (gToolbarDirtyFlags.has(BottomToolbarDirtyFlag::money))
-            {
-                gToolbarDirtyFlags.unset(BottomToolbarDirtyFlag::money);
-                invalidateWidget(WIDX_LEFT_INSET);
-            }
-
-            if (gToolbarDirtyFlags.has(BottomToolbarDirtyFlag::date))
-            {
-                gToolbarDirtyFlags.unset(BottomToolbarDirtyFlag::date);
-                invalidateWidget(WIDX_RIGHT_INSET);
-            }
-
-            if (gToolbarDirtyFlags.has(BottomToolbarDirtyFlag::guestCount))
-            {
-                gToolbarDirtyFlags.unset(BottomToolbarDirtyFlag::guestCount);
-                invalidateWidget(WIDX_LEFT_INSET);
-            }
-
-            if (gToolbarDirtyFlags.has(BottomToolbarDirtyFlag::weather))
-            {
-                gToolbarDirtyFlags.unset(BottomToolbarDirtyFlag::weather);
-                invalidateWidget(WIDX_RIGHT_INSET);
-            }
-
-            if (gToolbarDirtyFlags.has(BottomToolbarDirtyFlag::parkRating))
-            {
-                gToolbarDirtyFlags.unset(BottomToolbarDirtyFlag::parkRating);
-                invalidateWidget(WIDX_LEFT_INSET);
-            }
-        }
-
     public:
         GameBottomToolbar()
         {
-            setWidgets(window_game_bottom_toolbar_widgets);
+            setWidgets(kGameBottomToolbarWidgets);
 
             currentFrame = 0;
             initScrollWidgets();
@@ -443,17 +233,6 @@ namespace OpenRCT2::Ui::Windows
 
             switch (widgetIndex)
             {
-                case WIDX_LEFT_OUTSET:
-                case WIDX_MONEY:
-                    if (!getGameState().park.flags.has(ParkFlag::noMoney))
-                        ContextOpenWindow(WindowClass::finances);
-                    break;
-                case WIDX_GUESTS:
-                    ContextOpenWindowView(WindowView::parkGuests);
-                    break;
-                case WIDX_PARK_RATING:
-                    ContextOpenWindowView(WindowView::parkRating);
-                    break;
                 case WIDX_MIDDLE_INSET:
                     if (News::IsQueueEmpty())
                     {
@@ -485,29 +264,7 @@ namespace OpenRCT2::Ui::Windows
                             WindowScrollToLocation(*mainWindow, subjectLoc.value());
                     }
                     break;
-                case WIDX_RIGHT_OUTSET:
-                case WIDX_DATE:
-                    ContextOpenWindow(WindowClass::recentNews);
-                    break;
             }
-        }
-
-        StringWithArgs onTooltip(WidgetIndex widgetIndex, StringId fallback) override
-        {
-            const auto& gameState = getGameState();
-            auto ft = Formatter();
-
-            switch (widgetIndex)
-            {
-                case WIDX_MONEY:
-                    ft.Add<money64>(gameState.park.currentProfit);
-                    ft.Add<money64>(gameState.park.value);
-                    break;
-                case WIDX_PARK_RATING:
-                    ft.Add<int16_t>(gameState.park.rating);
-                    break;
-            }
-            return { fallback, ft };
         }
 
         void onPrepareDraw() override
@@ -520,46 +277,12 @@ namespace OpenRCT2::Ui::Windows
             windowPos.y = ContextGetHeight() - height;
 
             // Change height of widgets in accordance with line height.
-            widgets[WIDX_LEFT_OUTSET].bottom = widgets[WIDX_MIDDLE_OUTSET].bottom = widgets[WIDX_RIGHT_OUTSET].bottom
-                = line_height * 3 + 3;
-            widgets[WIDX_LEFT_INSET].bottom = widgets[WIDX_MIDDLE_INSET].bottom = widgets[WIDX_RIGHT_INSET].bottom = line_height
-                    * 3
-                + 1;
-
-            // Reposition left widgets in accordance with line height... depending on whether there is money in play.
-            if (getGameState().park.flags.has(ParkFlag::noMoney))
-            {
-                widgets[WIDX_MONEY].setHidden();
-                widgets[WIDX_GUESTS].top = 1;
-                widgets[WIDX_GUESTS].bottom = line_height + 7;
-                widgets[WIDX_PARK_RATING].top = line_height + 8;
-                widgets[WIDX_PARK_RATING].bottom = height - 1;
-            }
-            else
-            {
-                widgets[WIDX_MONEY].setVisible();
-                widgets[WIDX_MONEY].bottom = widgets[WIDX_MONEY].top + line_height;
-                widgets[WIDX_GUESTS].top = widgets[WIDX_MONEY].bottom + 1;
-                widgets[WIDX_GUESTS].bottom = widgets[WIDX_GUESTS].top + line_height;
-                widgets[WIDX_PARK_RATING].top = widgets[WIDX_GUESTS].bottom - 1;
-                widgets[WIDX_PARK_RATING].bottom = height - 1;
-            }
-
-            // Reposition right widgets in accordance with line height, too.
-            widgets[WIDX_DATE].bottom = line_height + 1;
+            widgets[WIDX_MIDDLE_OUTSET].bottom = line_height * 3 + 3;
+            widgets[WIDX_MIDDLE_INSET].bottom = line_height * 3 + 1;
 
             // Anchor the middle and right panel to the right
-            int32_t x = ContextGetWidth();
-            width = x;
-            x--;
-            widgets[WIDX_RIGHT_OUTSET].right = x;
-            x -= 2;
-            widgets[WIDX_RIGHT_INSET].right = x;
-            x -= 137;
-            widgets[WIDX_RIGHT_INSET].left = x;
-            x -= 2;
-            widgets[WIDX_RIGHT_OUTSET].left = x;
-            x--;
+            const auto width = ContextGetWidth();
+            auto x = width - 1 - 2 - 137 - 2 - 1;
             widgets[WIDX_MIDDLE_OUTSET].right = x;
             x -= 2;
             widgets[WIDX_MIDDLE_INSET].right = x;
@@ -567,8 +290,6 @@ namespace OpenRCT2::Ui::Windows
             widgets[WIDX_NEWS_LOCATE].right = x;
             x -= 23;
             widgets[WIDX_NEWS_LOCATE].left = x;
-            widgets[WIDX_DATE].left = widgets[WIDX_RIGHT_OUTSET].left + 2;
-            widgets[WIDX_DATE].right = widgets[WIDX_RIGHT_OUTSET].right - 2;
 
             if (News::IsQueueEmpty())
             {
@@ -618,31 +339,17 @@ namespace OpenRCT2::Ui::Windows
 
         void onDraw(RenderTarget& rt) override
         {
-            const auto& leftWidget = widgets[WIDX_LEFT_OUTSET];
-            const auto& rightWidget = widgets[WIDX_RIGHT_OUTSET];
             const auto& middleWidget = widgets[WIDX_MIDDLE_OUTSET];
-
-            // Draw panel grey backgrounds
-            auto leftTop = windowPos + ScreenCoordsXY{ leftWidget.left, leftWidget.top };
-            auto rightBottom = windowPos + ScreenCoordsXY{ leftWidget.right, leftWidget.bottom };
-            Rectangle::filter(rt, { leftTop, rightBottom }, FilterPaletteID::palette51);
-
-            leftTop = windowPos + ScreenCoordsXY{ rightWidget.left, rightWidget.top };
-            rightBottom = windowPos + ScreenCoordsXY{ rightWidget.right, rightWidget.bottom };
-            Rectangle::filter(rt, { leftTop, rightBottom }, FilterPaletteID::palette51);
 
             if (ThemeGetFlags() & UITHEME_FLAG_USE_FULL_BOTTOM_TOOLBAR)
             {
                 // Draw grey background
-                leftTop = windowPos + ScreenCoordsXY{ middleWidget.left, middleWidget.top };
-                rightBottom = windowPos + ScreenCoordsXY{ middleWidget.right, middleWidget.bottom };
+                auto leftTop = windowPos + ScreenCoordsXY{ middleWidget.left, middleWidget.top };
+                auto rightBottom = windowPos + ScreenCoordsXY{ middleWidget.right, middleWidget.bottom };
                 Rectangle::filter(rt, { leftTop, rightBottom }, FilterPaletteID::palette51);
             }
 
             drawWidgets(rt);
-
-            DrawLeftPanel(rt);
-            DrawRightPanel(rt);
 
             if (!News::IsQueueEmpty())
             {
@@ -659,27 +366,6 @@ namespace OpenRCT2::Ui::Windows
             currentFrame++;
             if (currentFrame >= 24)
                 currentFrame = 0;
-
-            InvalidateDirtyWidgets();
-        }
-
-        CursorID onCursor(WidgetIndex widgetIndex, const ScreenCoordsXY& screenCoords, CursorID cursorId) override
-        {
-            switch (widgetIndex)
-            {
-                case WIDX_MONEY:
-                case WIDX_GUESTS:
-                case WIDX_PARK_RATING:
-                case WIDX_DATE:
-                    gTooltipCloseTimeout = gCurrentRealTimeTicks + 2000;
-                    break;
-            }
-            return cursorId;
-        }
-
-        void onPeriodicUpdate() override
-        {
-            InvalidateDirtyWidgets();
         }
     };
