+++ src/openrct2-ui/windows/Themes.cpp
@@ -154,6 +154,8 @@ namespace OpenRCT2::Ui::Windows
     static WindowClass window_themes_tab_1_classes[] = {
         WindowClass::topToolbar,
         WindowClass::bottomToolbar,
+        WindowClass::parkInfoPanel,
+        WindowClass::dateInfoPanel,
         WindowClass::editorScenarioBottomToolbar,
         WindowClass::editorTrackBottomToolbar,
         WindowClass::titleMenu,
