+++ src/openrct2/interface/WindowClasses.h
@@ -18,11 +18,13 @@ namespace OpenRCT2
         mainWindow = 0,
         topToolbar = 1,
         bottomToolbar = 2,
+        parkInfoPanel = 3,
+        dateInfoPanel = 4,
         tooltip = 5,
         dropdown = 6,
         about = 8,
-        // WC_PUBLISHER_CREDITS = 9,
-        // WC_MUSIC_CREDITS = 10,
+        // publisherCredits = 9,
+        // musicCredits = 10,
         error = 11,
         ride = 12,
         rideConstruction = 13,
